requirements.txt
discord.py
aiohttp
dblpy